background task works - it periodocally downloads mails but no filtering is applied
