+++ Gmail_to_Trello/organizeGmailToTrello/myGmail/templates/myGmail/index.html
@@ -11,14 +11,7 @@
     <p> Select <a href="/boards">destination board and a keyword</a></p>
     <!--<p>  <a href="/send_mails_to_trello">Send mails to trello</a></p>-->
     <p><a href="/check_mails">Turn mail checking on</a></p>
-    <form action="/turn_on_off/" method="POST">
-        {% csrf_token %}
-      {{ form.non_field_errors }}
-      {{ form.as_p }}
-      <input id="status_on_off" type="text" name="status_on_off" hidden value="ON">
 
-      <input type="submit" value="Turn on off" name="submit">
-    </form>
 
 
   {% else %}
