+++ Gmail_to_Trello/organizeGmailToTrello/myGmail/templates/myGmail/view-boards.html
@@ -56,18 +56,6 @@
 
 </script>
 
-<script>
-      function pass_ids() {
-      var board_id = document.getElementById("board_id").value;
-      var list_id = document.getElementById("list_id").value;
-      console.log('value:');
-      console.log(board_id);
-      console.log(list_id);
-      document.getElementById('board_id_hidden').value = board_id
-      }
-
-</script>
-
 <div>
     Go to <a href="/index">main page</a>
