+++ Gmail_to_Trello/organizeGmailToTrello/organizeGmailToTrello/settings.py
@@ -39,6 +39,7 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'django.contrib.sites',
     'myGmail',
+    'background_task',
 
     'allauth',
     'allauth.account',
