+++ Gmail_to_Trello/organizeGmailToTrello/trello_destination.csv
@@ -10,3 +10,6 @@ ksi¹¿ka,605f39c01f1da512b61db21a
 odpoczynek,605f39c01f1da512b61db21a
 tuesday,605a20f510b27935090c1a0c
 something,605a20f510b27935090c1a0c
+potato,6059277bb44a1213f4ae3014
+dead,605a1fd9fdd0344df312bedd
+góry,605f39c01f1da512b61db21a
